refactoring hook_on_model_delete
