+++ lib/active_model_cachers.rb
@@ -3,7 +3,7 @@ require 'active_model_cachers/config'
 require 'active_model_cachers/cache_service_factory'
 require 'active_model_cachers/cacher'
 require 'active_model_cachers/hook_dependencies'
-require 'active_model_cachers/hook_model_delete'
+require 'active_model_cachers/hook_on_model_delete'
 require 'active_record'
 require 'active_record/relation'
