--- lib/active_model_cachers/hook_model_delete.rb
@@ -1,23 +1,0 @@
-require 'active_record'
-
-class ActiveRecord::Base
-  class << self
-    def on_delete(&callback)
-      delete_hooks << callback
-    end
-
-    def delete_hooks
-      @delete_hooks ||= []
-    end
-
-    def clear_delete_hooks
-      @delete_hooks = []
-    end
-
-    alias_method :delete_without_hook_by_active_record_cacher, :delete if not method_defined?(:delete_without_hook_by_active_record_cacher)
-    def delete(id)
-      delete_hooks.each{|s| s.call(id) }
-      delete_without_hook_by_active_record_cacher(id)
-    end
-  end
-end
